combine two helper methods into one
